Add logout and fixes ui bugs

+++ cmd/server/init.go
@@ -14,19 +14,28 @@ import (
 
 const (
 	sysGroupCode        = "sys_group"
+	sysGuestGroupCode   = "sys_guest"
 	sysGroupDescription = "internal system group"
+	sysGuestDescription = "internal system group for observing"
 	sysRouteTag         = "sys"
 )
 
-func resolveGroup(ctx context.Context, rep group.Repository) (*group.Group, error) {
+type Option struct {
+	Force bool
+}
 
-	g, err := rep.GetByCodeDb(ctx, sysGroupCode)
+func resolveGroup(ctx context.Context, rep group.Repository, o *Option) (*group.Group, error) {
+
+	g, err := rep.GetByCode(ctx, sysGroupCode)
 	if err != nil {
 		if err != group.ErrEntityNotFound {
 			return nil, err
 		}
 	}
 	if g != nil {
+		if !o.Force {
+			return g, nil
+		}
 		err = rep.Delete(ctx, g.Id)
 		if err != nil {
 			return nil, err
@@ -37,10 +46,37 @@ func resolveGroup(ctx context.Context, rep group.Repository) (*group.Group, erro
 	if err != nil {
 		return nil, err
 	}
+
 	return g, nil
 }
 
-func resolveUser(ctx context.Context, rep user.Repository) (*user.User, error) {
+func resolveGuestGroup(ctx context.Context, rep group.Repository, baseGroupId int, o *Option) (*group.Group, error) {
+
+	g, err := rep.GetByCode(ctx, sysGuestGroupCode)
+	if err != nil {
+		if err != group.ErrEntityNotFound {
+			return nil, err
+		}
+	}
+	if g != nil {
+		if !o.Force {
+			return g, nil
+		}
+		err = rep.Delete(ctx, g.Id)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	g, err = rep.InsertGroupBasedOnAnother(ctx, &group.Group{Code: sysGuestGroupCode, Description: sysGuestDescription}, baseGroupId)
+	if err != nil {
+		return nil, err
+	}
+
+	return g, nil
+}
+
+func resolveUser(ctx context.Context, rep user.Repository, o *Option) (*user.User, error) {
 	var login = viper.GetString("user.login")
 
 	u, err := rep.GetByExternalId(ctx, login)
@@ -50,6 +86,9 @@ func resolveUser(ctx context.Context, rep user.Repository) (*user.User, error) {
 		}
 	}
 	if u != nil {
+		if !o.Force {
+			return u, nil
+		}
 		err = rep.Delete(ctx, u.Id)
 		if err != nil {
 			return nil, err
@@ -63,8 +102,10 @@ func resolveUser(ctx context.Context, rep user.Repository) (*user.User, error) {
 	return u, nil
 }
 
-func resolveUserAndGroup(ctx context.Context, rep usergroup.Repository, userId, groupId int) error {
-	_, err := rep.Insert(ctx, []usergroup.Pair{{GroupId: groupId, UserId: userId}})
+func resolveUserAndGroup(ctx context.Context, rep usergroup.Repository, userId, groupId int, o *Option) error {
+	// todo: add option ..
+
+	_, err := rep.InsertMany(ctx, []usergroup.Pair{{GroupId: groupId, UserId: userId}})
 	return err
 }
 
@@ -85,7 +126,7 @@ func buildRoutes(r chi.Router) []route.Route {
 	return rs
 }
 
-func resolveRoutes(ctx context.Context, rep route.Repository, rs []route.Route) ([]route.Route, error) {
+func resolveRoutes(ctx context.Context, rep route.Repository, rs []route.Route, o *Option) ([]route.Route, error) {
 
 	out := make([]route.Route, 0, len(rs))
 
@@ -99,6 +140,10 @@ func resolveRoutes(ctx context.Context, rep route.Repository, rs []route.Route) 
 		}
 
 		if rr != nil {
+			if !o.Force {
+				out = append(out, rr.Route)
+				continue
+			}
 			err := rep.Delete(ctx, rr.Id)
 			if err != nil {
 				return nil, err
@@ -115,8 +160,11 @@ func resolveRoutes(ctx context.Context, rep route.Repository, rs []route.Route) 
 	return out, nil
 }
 
-func resolveGroupAndRoutes(ctx context.Context, rep grouproute.Repository, rs []route.Route, groupId int) error {
+func resolveGroupAndRoutes(ctx context.Context, rep grouproute.Repository, rs []route.Route, groupId int, o *Option) error {
 	groupRoutePairs := make([]grouproute.Pair, 0)
+
+	// todo: add options ...
+
 	for _, r := range rs {
 		groupRoutePairs = append(groupRoutePairs, grouproute.Pair{
 			GroupId: groupId,
